+++ docs/metro/js/metro.js
@@ -15586,15 +15586,20 @@ var Spinner = {
         maxValue: null,
         fixed: 0,
         repeatThreshold: 500,
+        hideCursor: false,
         clsSpinner: "",
-        clsSpinnerValue: "",
+        clsSpinnerInput: "",
         clsSpinnerButton: "",
         clsSpinnerButtonPlus: "",
         clsSpinnerButtonMinus: "",
+        onBeforeChange: Metro.noop_true,
         onChange: Metro.noop,
         onPlusClick: Metro.noop,
         onMinusClick: Metro.noop,
+        onArrowUp: Metro.noop,
+        onArrowDown: Metro.noop,
         onButtonClick: Metro.noop,
+        onArrowClick: Metro.noop,
         onSpinnerCreate: Metro.noop
     },
 
@@ -15624,28 +15629,27 @@ var Spinner = {
     _createStructure: function(){
         var element = this.element, o = this.options;
         var spinner = $("<div>").addClass("spinner").addClass("buttons-"+o.buttonsPosition).addClass(element[0].className).addClass(o.clsSpinner);
-        var wrapper = $("<div>").addClass("input-wrapper").addClass(o.clsSpinnerValue);
         var button_plus = $("<button>").attr("type", "button").addClass("button spinner-button spinner-button-plus").addClass(o.clsSpinnerButton + " " + o.clsSpinnerButtonPlus).html(o.plusIcon);
         var button_minus = $("<button>").attr("type", "button").addClass("button spinner-button spinner-button-minus").addClass(o.clsSpinnerButton + " " + o.clsSpinnerButtonMinus).html(o.minusIcon);
         var init_value = element.val().trim();
 
         if (!Utils.isValue(init_value)) {
-            init_value = 0;
             element.val(0);
         }
 
         element[0].className = '';
 
         spinner.insertBefore(element);
-        element.appendTo(spinner);
+        element.appendTo(spinner).addClass(o.clsSpinnerInput);
 
         element.addClass("original-input");
 
-        wrapper.appendTo(spinner);
         button_plus.appendTo(spinner);
         button_minus.appendTo(spinner);
 
-        wrapper.text(Number(init_value));
+        if (o.hideCursor === true) {
+            spinner.addClass("hide-cursor");
+        }
 
         if (o.disabled === true || element.is(":disabled")) {
             this.disable();
@@ -15658,9 +15662,7 @@ var Spinner = {
         var that = this, element = this.element, o = this.options;
         var spinner = element.closest(".spinner");
 
-        var spinnerButtonClick = function(context, threshold){
-            var button = $(context);
-            var plus = button.hasClass("spinner-button-plus");
+        var spinnerButtonClick = function(plus, threshold){
             var curr = element.val();
 
             var val = Number(element.val());
@@ -15675,29 +15677,44 @@ var Spinner = {
             that._setValue(val.toFixed(o.fixed), true);
 
             Utils.exec(plus ? o.onPlusClick : o.onMinusClick, [curr, val, element.val()], element[0]);
+            Utils.exec(plus ? o.onArrowUp : o.onArrowDown, [curr, val, element.val()], element[0]);
             Utils.exec(o.onButtonClick, [curr, val, element.val(), plus ? 'plus' : 'minus'], element[0]);
+            Utils.exec(o.onArrowClick, [curr, val, element.val(), plus ? 'plus' : 'minus'], element[0]);
 
             setTimeout(function(){
                 if (that.repeat_timer) {
-                    spinnerButtonClick(context, 100);
+                    spinnerButtonClick(plus, 100);
                 }
             }, threshold);
         };
 
         spinner.on(Metro.events.start, ".spinner-button", function(){
             that.repeat_timer = true;
-            spinnerButtonClick(this, o.repeatThreshold);
+            spinnerButtonClick($(this).hasClass("spinner-button-plus"), o.repeatThreshold);
         });
 
         spinner.on(Metro.events.stop, ".spinner-button", function(){
+            that.repeat_timer = false;
+        });
+
+        element.on(Metro.events.keydown, function(e){
+            if (e.keyCode === Metro.keyCode.UP_ARROW || e.keyCode === Metro.keyCode.DOWN_ARROW) {
+                that.repeat_timer = true;
+                spinnerButtonClick(e.keyCode === Metro.keyCode.UP_ARROW, o.repeatThreshold);
+            }
+        });
+
+        spinner.on(Metro.events.keyup, function(){
             that.repeat_timer = false;
         });
     },
 
     _setValue: function(val, trigger_change){
         var element = this.element, o = this.options;
-        var spinner = element.closest(".spinner");
-        var wrapper = spinner.find(".input-wrapper");
+
+        if (Utils.exec(o.onBeforeChange, [val], element[0]) !== true) {
+            return ;
+        }
 
         if (Utils.isValue(o.maxValue) && val > Number(o.maxValue)) {
             val =  Number(o.maxValue);
@@ -15708,7 +15725,6 @@ var Spinner = {
         }
 
         element.val(val);
-        wrapper.text(val);
 
         Utils.exec(o.onChange, [val], element[0]);