+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 + [x] Input: add attribute `data-cls-custom-button`
 + [x] Tag input: add observing attribute `value`
 + [x] Tag input: fix method `val()`
++ [x] Spinner: add events `onArrowUp`, `onArrowDown`, `onArrowClick`
 
 ### 4.2.20
 + [x] Table: fix default padding for `th` and `td`